requirements.txt
fastapi
uvicorn[standard]
celery[redis]
firebase-admin>=6.6.0 
python-dotenv
openai
pydantic-settings
google-cloud-firestore
google-cloud-pubsub
